Update GDrive link, fix bug
This fixes a bug in the logic that controlled processing cards.
The project incorrectly checked the number of available cards after
processing each card, which overwrote the value of `%cards_available`
each time. This check has been removed, as well as `%cards_available`,
and replaced with `%cards_complete`, which correctly tracks how
many cards have been "completed": This is set to 2 when there are no
cards available to process.

It also updates the Google Drive link to be the folder in which the
XML file is stored to prevent the link from having to be updated
each time this project is updated.